v.0.1.221023

+++ app/Http/Controllers/UserController.php
@@ -91,7 +91,7 @@ class UserController extends Controller
 
         if ($request->hasFile('photo')) {
             $path = 'photo/'.$update->photo;
-            if (File::exists($path)) {
+            if (File::exists($path)&& $path!='photo/default.jpg') {
                 File::delete($path);
             }
